+++ lib/src/stdlib/io.dart
@@ -2,6 +2,7 @@ import 'package:d4rt/src/environment.dart';
 import 'package:d4rt/src/stdlib/io/directory.dart';
 import 'package:d4rt/src/stdlib/io/file_system_entity.dart';
 import 'package:d4rt/src/stdlib/io/file.dart';
+import 'package:d4rt/src/stdlib/io/link.dart';
 import 'package:d4rt/src/stdlib/io/stdio.dart';
 import 'package:d4rt/src/stdlib/io/http.dart';
 import 'package:d4rt/src/stdlib/io/process.dart';
@@ -14,6 +15,7 @@ export 'package:d4rt/src/environment.dart';
 export 'package:d4rt/src/stdlib/io/directory.dart';
 export 'package:d4rt/src/stdlib/io/file_system_entity.dart';
 export 'package:d4rt/src/stdlib/io/file.dart';
+export 'package:d4rt/src/stdlib/io/link.dart';
 export 'package:d4rt/src/stdlib/io/stdio.dart';
 export 'package:d4rt/src/stdlib/io/http.dart';
 export 'package:d4rt/src/stdlib/io/process.dart';
@@ -24,13 +26,14 @@ export 'package:d4rt/src/stdlib/io/socket.dart';
 
 class IoStdlib {
   static void register(Environment environment) {
-    // Register FileSystemEntity classes (converted)
+    // Register FileSystemEntity classes (concrete first, base last)
+    environment.defineBridge(FileIo.definition);
+    environment.defineBridge(DirectoryIo.definition);
+    environment.defineBridge(LinkIo.definition);
     environment.defineBridge(FileSystemEntityIo.definition);
     environment.defineBridge(FileStatIo.definition);
     environment.defineBridge(FileSystemEntityTypeIo.definition);
     environment.defineBridge(FileSystemEventIo.definition);
-    environment.defineBridge(DirectoryIo.definition);
-    environment.defineBridge(FileIo.definition);
 
     // Register File-related classes
     environment.defineBridge(FileModeIo.definition);